Token refreshing and JWT-extended

+++ Section11/resources/item.py
@@ -1,67 +1,92 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import (
+    jwt_required,
+    get_jwt_claims,
+    get_jwt_identity,
+    jwt_optional,
+    fresh_jwt_required,
+)
 from models.item import ItemModel
 
-class Items(Resource):
-    
+
+class Item(Resource):
     parser = reqparse.RequestParser()
-    parser.add_argument('price', type=float, required=True
-                            , help="This field cannot be left blank!")
-    parser.add_argument('store_id', type=float, required=True
-                            , help="Item needs a store ID!")
-        
+    parser.add_argument(
+        "price", type=float, required=True, help="This field cannot be left blank!"
+    )
+    parser.add_argument(
+        "store_id", type=int, required=True, help="Every item needs a store_id."
+    )
+
     @jwt_required()
     def get(self, name):
-        """GET method implementation"""
-        
         item = ItemModel.find_by_name(name)
-        
         if item:
-            return item.json(), 200
-    
-        return {'message': 'Item not found'}, 404
-          
+            return item.json()
+        return {"message": "Item not found"}, 404
+
+    @jwt_required(fresh=True)
     def post(self, name):
-        """POST method implementation"""
-               
         if ItemModel.find_by_name(name):
-            return {'message': 'Item with name {} already exists'.format(name)}, 400
-        
-        data = self.parser.parse_args()
-        new_item = ItemModel(name, data['price'], data['store_id'])
-        
+            return {
+                "message": "An item with name '{}' already exists.".format(name)
+            }, 400
+
+        data = Item.parser.parse_args()
+
+        item = ItemModel(name, **data)
+
         try:
-            new_item.save_to_db()
+            item.save_to_db()
         except:
-            return {'message': "An error occured inserting the item"}, 500  #internal server error
-        
-        return new_item.json(), 201
-    
+            return {"message": "An error occurred inserting the item."}, 500
+
+        return item.json(), 201
+
+    @jwt_required()
     def delete(self, name):
-        
+        claims = get_jwt_claims()
+        if not claims["is_admin"]:
+            return {"message": "Admin privilege required."}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
-            
-        return {'message': 'Item deleted'}
-    
+            return {"message": "Item deleted."}
+        return {"message": "Item not found."}, 404
+
     def put(self, name):
-              
-        data = self.parser.parse_args()
+        data = Item.parser.parse_args()
+
         item = ItemModel.find_by_name(name)
-        
-        if not item:
-            item = ItemModel(name, data['price'], data['store_id'])
+
+        if item:
+            item.price = data["price"]
         else:
-            item.price = data['price']
-        
+            item = ItemModel(name, **data)
+
         item.save_to_db()
-        
+
         return item.json()
 
-        
-class itemList(Resource):
+
+class ItemList(Resource):
+    @jwt_required(optional=True)
     def get(self):
-        """Get item list method"""
-        return {'items': [item.json() for item in ItemModel.query.all()]}
-    
+        """
+        Here we get the JWT identity, and then if the user is logged in (we were able to get an identity)
+        we return the entire item list.
+
+        Otherwise we just return the item names.
+
+        This could be done with e.g. see orders that have been placed, but not see details about the orders
+        unless the user has logged in.
+        """
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {"items": items}, 200
+        return {
+            "items": [item["name"] for item in items],
+            "message": "More data available if you log in.",
+        }, 200
